requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community>=0.0.28,<0.1
python-dotenv==1.0.1
typing-extensions==4.9.0
requests==2.31.0
python-magic==0.4.27
Pillow==10.4.0
pdf2image==1.17.0
pytesseract==0.3.10
python-docx==1.1.0
PyPDF2==3.0.1
unstructured==0.11.8
psycopg2-binary==2.9.9
pgvector==0.2.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
cryptography==44.0.3
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.1.post1
sentence-transformers==2.5.1
opencv-python-headless==4.9.0.80
pdfminer.six==20231228
python-magic-bin==0.4.14; platform_system == "Windows"
poppler-utils==0.1.0
tesseract==0.1.3
psutil==5.9.8
watchdog==4.0.0 